feat: allow to submit two or more vouchers
